pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ethos-py"
version = "0.1.0"
description = "Unofficial Python SDK for Ethos Network API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "onetrillionx", email = "onetrillionx@proton.me" }
]
keywords = [
    "ethos",
    "ethos-network",
    "web3",
    "reputation",
    "blockchain",
    "api",
    "sdk",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/onetrillionx/ethos-python-sdk"
Documentation = "https://github.com/onetrillionx/ethos-python-sdk#readme"
Repository = "https://github.com/onetrillionx/ethos-python-sdk"
Issues = "https://github.com/onetrillionx/ethos-python-sdk/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ethos"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["ethos"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["_pytest"]
# Suppress false positives where mypy confuses method names with types
disable_error_code = ["valid-type"]

[[tool.mypy.overrides]]
module = "ethos.types.*"
# Pydantic models need flexibility
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "_pytest.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/ethos --cov-report=term-missing"

[tool.coverage.run]
source = ["src/ethos"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
